+++ src/components/sets/reducer.test.js
@@ -1,28 +1,21 @@
-/* eslint-disable */
+/* eslint-disable no-underscore-dangle */
 import expect from 'expect';
 import reducer, {
-  treeInitialize, treeExport, treeToVisibleCells, nodeToRenderProps, ACTION,
+  treeInitialize, nodeToRenderProps, ACTION,
 } from './reducer';
 
 import {
-  levelTwoNodeLeafWithoutState,
   levelTwoNodeLeaf,
-  levelOneNodeWithoutState,
-  levelOneNode,
-  levelZeroNodeWithoutState,
   levelZeroNode,
   treeWithoutState,
   treeWithoutStateOrColors,
   treeIgnoreKeys,
   tree,
-  emptyTreeWithoutState,
-  emptyTreeIgnoreKeys,
   emptyTree,
 } from './reducer.test.fixtures';
 
 
 describe('Hierarchical sets reducer', () => {
-
   describe('Node rendering', () => {
     it('can get render properties for a node', () => {
       const levelTwoRenderProps = nodeToRenderProps(levelTwoNodeLeaf);
@@ -51,10 +44,9 @@ describe('Hierarchical sets reducer', () => {
       expect(levelZeroRenderProps.isLeaf).toEqual(false);
       expect(levelZeroRenderProps.height).toEqual(2);
     });
-  })
+  });
 
   describe('Tree manipulation', () => {
-
     it('can be initialized', () => {
       const initialTree = treeInitialize('cell');
 
@@ -81,7 +73,7 @@ describe('Hierarchical sets reducer', () => {
 
       expect(reducer(initialTree, {
         type: ACTION.IMPORT,
-        levelZeroNodes: treeWithoutState.tree
+        levelZeroNodes: treeWithoutState.tree,
       })).toMatchObject(treeIgnoreKeys);
     });
 
@@ -90,7 +82,7 @@ describe('Hierarchical sets reducer', () => {
 
       const postImportTree = reducer(initialTree, {
         type: ACTION.IMPORT,
-        levelZeroNodes: treeWithoutStateOrColors.tree
+        levelZeroNodes: treeWithoutStateOrColors.tree,
       });
 
       expect(postImportTree.tree[0].color).toEqual(undefined);
@@ -102,7 +94,7 @@ describe('Hierarchical sets reducer', () => {
 
       const postImportTree = reducer(initialTree, {
         type: ACTION.IMPORT,
-        levelZeroNodes: treeWithoutState.tree
+        levelZeroNodes: treeWithoutState.tree,
       });
 
       const keySet = new Set();
@@ -111,8 +103,13 @@ describe('Hierarchical sets reducer', () => {
       keySet.add(postImportTree.tree[0].children[0]._state.key);
       keySet.add(postImportTree.tree[0].children[0].children[0]._state.key);
       keySet.add(postImportTree.tree[0].children[0].children[1]._state.key);
-
       expect(keySet.size).toEqual(5);
+
+      // Also check that it assigns the correct level.
+      expect(postImportTree.tree[0]._state.level).toEqual(0);
+      expect(postImportTree.tree[0].children[0]._state.level).toEqual(1);
+      expect(postImportTree.tree[0].children[0].children[0]._state.level).toEqual(2);
+      expect(postImportTree.tree[0].children[0].children[1]._state.level).toEqual(2);
     });
 
     it('can check leaf nodes', () => {
@@ -133,7 +130,7 @@ describe('Hierarchical sets reducer', () => {
       });
       expect(postSecondCheckTree._state.checkedKeys).toEqual([
         'vasculature-endothelial',
-        'vasculature-pericytes'
+        'vasculature-pericytes',
       ]);
     });
 
@@ -201,12 +198,41 @@ describe('Hierarchical sets reducer', () => {
       expect(postUncheckTree._state.checkedKeys).toEqual([]);
     });
 
-    it('can set the current set', () => {
-      
+    it('can set the current selection', () => {
+      const initialTree = tree;
+      expect(initialTree.tree.length).toEqual(1);
+      const postSelectionTree = reducer(initialTree, {
+        type: ACTION.SET_CURRENT_SET,
+        cellIds: ['cell_x', 'cell_y', 'cell_z'],
+      });
+      // Expect the current selection set to the the only one visible.
+      expect(postSelectionTree._state.visibleKeys.length).toEqual(1);
+      // Expect there to be one additional expanded set
+      // (corresponding to the .isForTools node).
+      expect(postSelectionTree._state.expandedKeys.length).toEqual(1);
+
+      // Expect there to be an additional level zero node.
+      expect(postSelectionTree.tree.length).toEqual(2);
+      expect(postSelectionTree.tree[1].name).toEqual('My Selections');
+      expect(postSelectionTree.tree[1]._state.isForTools).toEqual(true);
+      expect(postSelectionTree.tree[1]._state.level).toEqual(0);
+      expect(postSelectionTree.tree[0]._state.isForTools).toEqual(false);
+      expect(postSelectionTree.tree[1].children.length).toEqual(1);
+      expect(postSelectionTree.tree[1].children[0].name).toEqual('Current selection');
+      expect(postSelectionTree.tree[1].children[0]._state.isCurrent).toEqual(true);
+      expect(postSelectionTree.tree[1].children[0]._state.level).toEqual(1);
+      expect(postSelectionTree.tree[1].children[0].set).toEqual(['cell_x', 'cell_y', 'cell_z']);
     });
 
     it('can start editing a node', () => {
-     
+      const initialTree = tree;
+      const editingTree = reducer(initialTree, {
+        type: ACTION.SET_NODE_IS_EDITING,
+        targetKey: 'vasculature-endothelial',
+        value: true,
+      });
+      expect(editingTree.tree[0].children[0].children[0]._state.isEditing).toEqual(false);
+      expect(editingTree.tree[0].children[0].children[1]._state.isEditing).toEqual(true);
     });
 
     it('can delete a leaf node', () => {
@@ -231,7 +257,7 @@ describe('Hierarchical sets reducer', () => {
 
     it('can change a node name', () => {
       const initialTree = tree;
-      expect(initialTree.tree[0].children[0].name).toEqual("Vasculature");
+      expect(initialTree.tree[0].children[0].name).toEqual('Vasculature');
       const postRenameTree = reducer(initialTree, {
         type: ACTION.SET_NODE_NAME,
         targetKey: 'vasculature',
@@ -256,67 +282,145 @@ describe('Hierarchical sets reducer', () => {
       const postRecolorTree = reducer(initialTree, {
         type: ACTION.SET_NODE_COLOR,
         targetKey: 'vasculature',
-        color: [1, 2, 3]
+        color: [1, 2, 3],
       });
       expect(postRecolorTree.tree[0].children[0].color).toEqual([1, 2, 3]);
     });
 
     it('can view expanded descendant sets of a non-leaf node', () => {
-      
-    });
-
-    it('can do a union operation', () => {
-      
-    });
-
-    it('can do an intersection operation', () => {
-      
-    });
-
-    it('can do a complement operation', () => {
-      
-    });
-
-    it('can move a drag node to a drop node, making drag node the only child of drop node', () => {
-      
-    });
-
-    it('can move a drag node to a drop node, making drag node the last child of drop node', () => {
-      
-    });
-
-    it('can move a drag node up, below a drop node, into the gap between two nodes', () => {
-      
-    });
-
-    it('can move a drag node up, above a drop node, into the gap between two nodes', () => {
-      
-    });
-
-    it('can move a drag node up, above a drop node that is a first child', () => {
-      
-    });
-
-    it('can move a drag node down, above a drop node, into the gap between two nodes', () => {
-      
-    });
-
-    it('can move a drag node down, below a drop node that is a last child', () => {
-     
-    });
-
-    it('can move a drag node up a level, below its parent node', () => {
-     
-    });
-
-    it('can move a drag node up a level, above its parent node', () => {
-      
+      const initialTree = tree;
+      expect(initialTree._state.expandedKeys).toEqual([]);
+      const postExpandTree = reducer(initialTree, {
+        type: ACTION.EXPAND_NODE,
+        expandedKeys: ['cell-type-annotations'],
+        targetKey: 'cell-type-annotations',
+        expanded: true,
+      });
+      expect(postExpandTree._state.expandedKeys).toEqual(['cell-type-annotations']);
     });
 
     it('can view a leaf node set', () => {
-      
+      const initialTree = tree;
+      expect(initialTree._state.visibleKeys).toEqual([]);
+      const postViewTree = reducer(initialTree, {
+        type: ACTION.VIEW_NODE,
+        targetKey: 'vasculature-endothelial',
+      });
+      expect(postViewTree._state.visibleKeys).toEqual(['vasculature-endothelial']);
     });
 
+    it('can do a union operation', () => {
+      const initialTree = tree;
+      expect(initialTree.tree.length).toEqual(1);
+      const postCheckTree = reducer(initialTree, {
+        type: ACTION.CHECK_NODE,
+        targetKey: 'vasculature-endothelial',
+        checked: true,
+      });
+      const postSecondCheckTree = reducer(postCheckTree, {
+        type: ACTION.CHECK_NODE,
+        targetKey: 'vasculature-pericytes',
+        checked: true,
+      });
+      const postUnionTree = reducer(postSecondCheckTree, {
+        type: ACTION.UNION_CHECKED,
+      });
+      expect(postUnionTree.tree.length).toEqual(2);
+      expect(postUnionTree.tree[1].name).toEqual('My Selections');
+      expect(postUnionTree.tree[1]._state.isForTools).toEqual(true);
+      expect(postUnionTree.tree[1].children[0].name).toEqual('Current union');
+      expect(postUnionTree.tree[1].children[0]._state.isCurrent).toEqual(true);
+      expect(postUnionTree.tree[1].children[0]._state.level).toEqual(1);
+      expect(postUnionTree.tree[1].children[0].set).toEqual([
+        'cell_3', 'cell_4', 'cell_5', 'cell_1', 'cell_2',
+      ]);
+      expect(postUnionTree._state.visibleKeys.length).toEqual(1);
+    });
+
+    it('can do an intersection operation', () => {
+      const initialTree = tree;
+      expect(initialTree.tree.length).toEqual(1);
+      const postCheckTree = reducer(initialTree, {
+        type: ACTION.CHECK_NODE,
+        targetKey: 'vasculature-endothelial',
+        checked: true,
+      });
+      const postSecondCheckTree = reducer(postCheckTree, {
+        type: ACTION.CHECK_NODE,
+        targetKey: 'vasculature-pericytes',
+        checked: true,
+      });
+      const postIntersectionTree = reducer(postSecondCheckTree, {
+        type: ACTION.INTERSECTION_CHECKED,
+      });
+      expect(postIntersectionTree.tree.length).toEqual(2);
+      expect(postIntersectionTree.tree[1].name).toEqual('My Selections');
+      expect(postIntersectionTree.tree[1]._state.isForTools).toEqual(true);
+      expect(postIntersectionTree.tree[1].children[0].name).toEqual('Current intersection');
+      expect(postIntersectionTree.tree[1].children[0]._state.isCurrent).toEqual(true);
+      expect(postIntersectionTree.tree[1].children[0]._state.level).toEqual(1);
+      expect(postIntersectionTree.tree[1].children[0].set).toEqual([
+        'cell_3',
+      ]);
+      expect(postIntersectionTree._state.visibleKeys.length).toEqual(1);
+    });
+
+    it('can do a complement operation', () => {
+      const initialTree = tree;
+      expect(initialTree.tree.length).toEqual(1);
+      const postCheckTree = reducer(initialTree, {
+        type: ACTION.CHECK_NODE,
+        targetKey: 'vasculature-endothelial',
+        checked: true,
+      });
+      const postSecondCheckTree = reducer(postCheckTree, {
+        type: ACTION.CHECK_NODE,
+        targetKey: 'vasculature-pericytes',
+        checked: true,
+      });
+      const postIntersectionTree = reducer(postSecondCheckTree, {
+        type: ACTION.COMPLEMENT_CHECKED,
+      });
+      expect(postIntersectionTree.tree.length).toEqual(2);
+      expect(postIntersectionTree.tree[1].name).toEqual('My Selections');
+      expect(postIntersectionTree.tree[1]._state.isForTools).toEqual(true);
+      expect(postIntersectionTree.tree[1].children[0].name).toEqual('Current complement');
+      expect(postIntersectionTree.tree[1].children[0]._state.isCurrent).toEqual(true);
+      expect(postIntersectionTree.tree[1].children[0]._state.level).toEqual(1);
+      expect(postIntersectionTree.tree[1].children[0].set).toEqual([
+        'cell_6',
+      ]);
+      expect(postIntersectionTree._state.visibleKeys.length).toEqual(1);
+    });
+
+    it('can drag to rearrange leaf nodes, dragging to top of list', () => {
+      const initialTree = tree;
+      expect(initialTree.tree[0].children[0].children[0]._state.key).toEqual('vasculature-pericytes');
+      expect(initialTree.tree[0].children[0].children[1]._state.key).toEqual('vasculature-endothelial');
+      const postDragNode = reducer(initialTree, {
+        type: ACTION.DROP_NODE,
+        dropKey: 'vasculature-pericytes',
+        dragKey: 'vasculature-endothelial',
+        dropPosition: -1,
+        dropToGap: true,
+      });
+      expect(postDragNode.tree[0].children[0].children[0]._state.key).toEqual('vasculature-endothelial');
+      expect(postDragNode.tree[0].children[0].children[1]._state.key).toEqual('vasculature-pericytes');
+    });
+
+    it('can drag to rearrange leaf nodes, dragging to bottom of list', () => {
+      const initialTree = tree;
+      expect(initialTree.tree[0].children[0].children[0]._state.key).toEqual('vasculature-pericytes');
+      expect(initialTree.tree[0].children[0].children[1]._state.key).toEqual('vasculature-endothelial');
+      const postDragNode = reducer(initialTree, {
+        type: ACTION.DROP_NODE,
+        dropKey: 'vasculature-endothelial',
+        dragKey: 'vasculature-pericytes',
+        dropPosition: 2,
+        dropToGap: true,
+      });
+      expect(postDragNode.tree[0].children[0].children[0]._state.key).toEqual('vasculature-endothelial');
+      expect(postDragNode.tree[0].children[0].children[1]._state.key).toEqual('vasculature-pericytes');
+    });
   });
 });
-